Reorgd PyPi and Nuget pkg to use source from -Dev folder

+++ PyPi/CodeProject.AI-SDK/build.bat
@@ -1,3 +1,7 @@
 :: This file builds the package
-del /s /q dist
+
+if exist dist del /s /q dist
+
+robocopy /e "..\..\..\CodeProject.AI-Server-Dev\src\SDK\Python\src\codeproject_ai_sdk\ " ".\build\lib\codeproject_ai_sdk " > nul
+
 python -m build --wheel
